CT-5: Move `Artist` from base `/model` -> `/model/data`
