+++ src/main/kotlin/com/evanconell/concerttrackerapi/respository/ArtistRepository.kt
@@ -1,6 +1,6 @@
 package com.evanconell.concerttrackerapi.respository
 
-import com.evanconell.concerttrackerapi.model.Artist
+import com.evanconell.concerttrackerapi.model.data.Artist
 import org.springframework.data.repository.CrudRepository
 
 interface ArtistRepository : CrudRepository<Artist, String> {
